Rename Tournament module to TournamentSystem,
Tournament is too generic name.
