+++ lib/tournament-system/algorithm.rb
@@ -1,0 +1,6 @@
+module TournamentSystem
+  # This module provides abtraction-less implementations of all algorithms
+  # used for tournament systems.
+  module Algorithm
+  end
+end
