+++ lib/tournament-system/version.rb
@@ -1,0 +1,4 @@
+module TournamentSystem
+  # The current version of this gem.
+  VERSION = '1.0.2'.freeze
+end
