--- lib/tournament.rb
@@ -1,6 +1,0 @@
-# This library is split into two parts, there's the actual algorithms that
-# implement various tournament systems ({Algorithm}) and a data abstraction
-# layer for generating matches using various tournament systems in a
-# data-independent way ({Driver}).
-module Tournament
-end
