--- lib/tournament/algorithm.rb
@@ -1,6 +1,0 @@
-module Tournament
-  # This module provides abtraction-less implementations of all algorithms
-  # used for tournament systems.
-  module Algorithm
-  end
-end
