--- lib/tournament/version.rb
@@ -1,4 +1,0 @@
-module Tournament
-  # The current version of this gem.
-  VERSION = '1.0.2'.freeze
-end
